+++ src/qureg.rs
@@ -228,6 +228,183 @@ impl<'a, const N: usize> Qureg<'a, N> {
         })
         .expect("report_qureg_params should never fail");
     }
+
+    /// Initializes a `Qureg` to have all-zero-amplitudes.
+    ///
+    /// This is an unphysical state, useful for iteratively building a state
+    /// with functions like
+    /// [`set_weighted_qureg()`][api-set-weighted-qureg], and should
+    /// not be confused with [`init_zero_state()`][api-init-zero-state].
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// # use quest_bind::*;
+    /// let env = &QuestEnv::new();
+    /// let qureg = &mut Qureg::try_new(2, env).unwrap();
+    ///
+    /// init_blank_state(qureg);
+    ///
+    /// assert!(get_prob_amp(qureg, 0).unwrap().abs() < EPSILON);
+    /// assert!(get_prob_amp(qureg, 1).unwrap().abs() < EPSILON);
+    /// assert!(get_prob_amp(qureg, 2).unwrap().abs() < EPSILON);
+    /// assert!(get_prob_amp(qureg, 3).unwrap().abs() < EPSILON);
+    /// ```
+    ///
+    /// See [QuEST API] for more information.
+    ///
+    /// [api-set-weighted-qureg]: crate::set_weighted_qureg()
+    /// [api-init-zero-state]: crate::Qureg::init_zero_state()
+    /// [api-qureg]: crate::Qureg
+    /// [QuEST API]: https://quest-kit.github.io/QuEST/modules.html
+    #[allow(clippy::needless_pass_by_ref_mut)]
+    pub fn init_blank_state(&mut self) {
+        catch_quest_exception(|| unsafe {
+            ffi::initBlankState(self.reg);
+        })
+        .expect("init_blank_state should always succeed");
+    }
+
+    /// Initialize `qureg` into the zero state.
+    ///
+    /// If `qureg` is a state-vector of `N` qubits, it is modified to state
+    /// `|0>^{\otimes N}`.  If `qureg` is a density matrix of `N` qubits, it is
+    /// modified to state `|0><0|^{\otimes N}`.
+    ///
+    /// # Parameters
+    ///
+    /// - `qureg`: a [`Qureg`][api-qureg] of which to clear all amplitudes
+    ///
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// # use quest_bind::*;
+    /// let env = &QuestEnv::new();
+    /// let qureg = &mut Qureg::try_new(2, env).unwrap();
+    ///
+    /// init_zero_state(qureg);
+    ///
+    /// assert!((get_prob_amp(qureg, 0).unwrap() - 1.).abs() < EPSILON);
+    /// assert!(get_prob_amp(qureg, 1).unwrap().abs() < EPSILON);
+    /// assert!(get_prob_amp(qureg, 2).unwrap().abs() < EPSILON);
+    /// assert!(get_prob_amp(qureg, 3).unwrap().abs() < EPSILON);
+    /// ```
+    ///
+    /// See [QuEST API] for more information.
+    ///
+    /// [api-qureg]: crate::Qureg
+    /// [QuEST API]: https://quest-kit.github.io/QuEST/modules.html
+    #[allow(clippy::needless_pass_by_ref_mut)]
+    pub fn init_zero_state(&mut self) {
+        catch_quest_exception(|| unsafe {
+            ffi::initZeroState(self.reg);
+        })
+        .expect("init_zero_state should always succeed");
+    }
+
+    /// Initialize `qureg` into the plus state.
+    ///
+    /// If `qureg` is a state-vector of `N` qubits, it is modified to state:
+    ///
+    /// ```latex
+    ///   {| + \rangle}^{\otimes N} = \frac{1}{\sqrt{2^N}} (| 0 \rangle + | 1 \rangle)^{\otimes N}.
+    /// ```
+    ///
+    /// If `qureg` is a density matrix of `N`, it is modified to state:
+    ///
+    /// ```latex
+    ///   {| + \rangle\langle+|}^{\otimes N} = \frac{1}{{2^N}} \sum_i\sum_j |i\rangle\langle j|.
+    /// ```
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// # use quest_bind::*;
+    /// let env = &QuestEnv::new();
+    /// let qureg = &mut Qureg::try_new(2, env).unwrap();
+    ///
+    /// init_plus_state(qureg);
+    ///
+    /// assert!((get_prob_amp(qureg, 0).unwrap() - 0.25).abs() < EPSILON);
+    /// assert!((get_prob_amp(qureg, 1).unwrap() - 0.25).abs() < EPSILON);
+    /// assert!((get_prob_amp(qureg, 2).unwrap() - 0.25).abs() < EPSILON);
+    /// assert!((get_prob_amp(qureg, 3).unwrap() - 0.25).abs() < EPSILON);
+    /// ```
+    ///
+    /// See [QuEST API] for more information.
+    ///
+    /// [api-qureg]: crate::Qureg
+    /// [QuEST API]: https://quest-kit.github.io/QuEST/modules.html
+    #[allow(clippy::needless_pass_by_ref_mut)]
+    pub fn init_plus_state(&mut self) {
+        catch_quest_exception(|| unsafe {
+            ffi::initPlusState(self.reg);
+        })
+        .expect("init_plus_state should always succeed");
+    }
+
+    /// Initialize `qureg` into a classical state.
+    ///
+    /// This state is also known as a "computational basis state" with index
+    /// `state_ind`.
+    ///
+    /// If `qureg` is a state-vector, it will become: `|state_ind>`. If
+    /// `qureg`is a density matrix, it will become:
+    ///
+    /// ```text
+    ///   |state_ind> <state_ind|
+    /// ````
+    ///
+    /// Classical states are indexed from zero, so that `state_ind=0` produces
+    /// `|0..00>`,  and  `state_ind=1` produces `|00..01>`, and `state_ind=2^N -
+    /// 1` produces `|11..11>`. Subsequent calls to
+    /// [`get_prob_amp()`][api-get-prob-amp] will yield `0` for all indices
+    /// except `state_ind`,  and the phase of `state_ind`'s amplitude will
+    /// be `1` (real).
+    ///
+    /// This function can be used to initialise `qureg` into a specific binary
+    /// state  (e.g. `11001`) using a binary literal.
+    ///
+    /// # Parameters
+    ///
+    ///  - `qureg`: the register to modify
+    ///  - `state_ind` the index of the basis state to modify `qureg` into
+    ///
+    /// # Errors
+    ///
+    /// - [`InvalidQuESTInputError`],
+    ///   - if `state_ind` is outside [0, qureg.[`num_qubits_represented()`]).
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// # use quest_bind::*;
+    /// let env = &QuestEnv::new();
+    /// let qureg = &mut Qureg::try_new(3, env).unwrap();
+    ///
+    /// init_classical_state(qureg, 8);
+    /// let prob = get_prob_amp(qureg, 0).unwrap();
+    ///
+    /// assert!((prob.abs() - 1.) < EPSILON);
+    /// ```
+    ///
+    ///
+    /// See [QuEST API] for more information.
+    ///
+    /// [api-get-prob-amp]: crate::get_prob_amp()
+    /// [`num_qubits_represented()`]: crate::Qureg::num_qubits_represented()
+    /// [`InvalidQuESTInputError`]: crate::QuestError::InvalidQuESTInputError
+    /// [QuEST API]: https://quest-kit.github.io/QuEST/modules.html
+    #[allow(clippy::needless_pass_by_ref_mut)]
+    pub fn init_classical_state(
+        &mut self,
+        state_ind: i64,
+    ) -> Result<(), QuestError> {
+        catch_quest_exception(|| unsafe {
+            ffi::initClassicalState(self.reg, state_ind);
+        })
+    }
 }
 
 impl<'a, const N: usize> Drop for Qureg<'a, N> {